Render operator.

Vectorizes function whisker.render.

NOTE: Because whisker does not support variable names inlcuding dots, a dot in column names is replaced by an underscore. Therefore, you must use an underscore in the template text for varibales including dots.

template %$$% data

Arguments

template
template

data
data frame where column names correspond to variables names in template

Value

rendered template

See also

Examples

vars <- data.frame(a = 1:3, b = 23:21)
"a = {{a}} and b = {{b}}" %$$% vars

#> [1] "a = 1 and b = 23" "a = 2 and b = 22" "a = 3 and b = 21"